backend/requirements.txt
appwrite==4.1.0
python-dotenv==1.0.0
Flask==3.0.1
Flask-Cors==4.0.0
mercadopago==2.2.1
pymongo==4.6.2
cryptography==42.0.5